- updating to latest trunk (Aly)
git-svn-id: svn://svn.zabbix.com/branches/dev@11467 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ dev-449-host_time/frontends/php/api/classes/class.cuser.php
@@ -1,7 +1,7 @@
 <?php
 /*
 ** ZABBIX
-** Copyright (C) 2000-2009 SIA Zabbix
+** Copyright (C) 2000-2010 SIA Zabbix
 **
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
@@ -29,14 +29,6 @@
 class CUser extends CZBXAPI{
 /**
  * Get Users data
- *
- * First part of parameters are filters which limits the output result set, these filters are set only if appropriate parameter is set.
- * For example if "type" is set, then method returns only users of given type.
- * Second part of parameters extends result data, adding data about others objects that are related to objects we get.
- * For example if "select_usrgrps" parameter is set, resulting objects will have additional property 'usrgrps' containing object with
- * data about User UserGroups.
- * Third part of parameters affect output. For example "sortfield" will be set to 'alias', result will be sorted by User alias.
- * All Parameters are optional!
  *
  * {@source}
  * @access public
@@ -87,7 +79,7 @@ class CUser extends CZBXAPI{
 			'pattern'					=> '',
 // OutPut
 			'extendoutput'				=> null,
-			'output'				=> API_OUTPUT_REFER,
+			'output'					=> API_OUTPUT_REFER,
 			'editable'					=> null,
 			'select_usrgrps'			=> null,
 			'get_access'				=> null,
@@ -101,26 +93,35 @@ class CUser extends CZBXAPI{
 
 		$options = zbx_array_merge($def_options, $options);
 
-		
+
 		if(!is_null($options['extendoutput'])){
 			$options['output'] = API_OUTPUT_EXTEND;
-			
+
 			if(!is_null($options['select_usrgrps'])){
 				$options['select_usrgrps'] = API_OUTPUT_EXTEND;
 			}
 		}
-		
-		
+
+
 // PERMISSION CHECK
 		if(USER_TYPE_SUPER_ADMIN == $user_type){
 
+		}
+		else if(is_null($options['editable']) && ($USER_DETAILS['type'] == USER_TYPE_ZABBIX_ADMIN)){
+			$sql_parts['from']['ug'] = 'users_groups ug';
+			$sql_parts['where']['uug'] = 'u.userid=ug.userid';
+			$sql_parts['where'][] = 'ug.usrgrpid IN ('.
+				' SELECT uug.usrgrpid'.
+				' FROM users_groups uug'.
+				' WHERE uug.userid='.$USER_DETAILS['userid'].
+				' )';
 		}
 		else if(!is_null($options['editable']) && ($USER_DETAILS['type']!=USER_TYPE_SUPER_ADMIN)){
 			return array();
 		}
 
 // nodeids
-		$nodeids = $options['nodeids'] ? $options['nodeids'] : get_current_nodeid(false);
+		$nodeids = !is_null($options['nodeids']) ? $options['nodeids'] : get_current_nodeid(false);
 
 // usrgrpids
 		if(!is_null($options['usrgrpids'])){
@@ -166,7 +167,7 @@ class CUser extends CZBXAPI{
 
 // pattern
 		if(!zbx_empty($options['pattern'])){
-			$sql_parts['where'][] = ' UPPER(u.alias) LIKE '.zbx_dbstr('%'.strtoupper($options['pattern']).'%');
+			$sql_parts['where'][] = ' UPPER(u.alias) LIKE '.zbx_dbstr('%'.zbx_strtoupper($options['pattern']).'%');
 		}
 
 // order
@@ -204,7 +205,7 @@ class CUser extends CZBXAPI{
 		if(!empty($sql_parts['order']))		$sql_order.= ' ORDER BY '.implode(',',$sql_parts['order']);
 		$sql_limit = $sql_parts['limit'];
 
-		$sql = 'SELECT '.$sql_select.'
+		$sql = 'SELECT DISTINCT '.$sql_select.'
 				FROM '.$sql_from.'
 				WHERE '.DBin_node('u.userid', $nodeids).
 				$sql_where.
@@ -228,11 +229,11 @@ class CUser extends CZBXAPI{
 					}
 
 // usrgrpids
-					if(isset($user['usrgrpid'])){
+					if(isset($user['usrgrpid'])  && is_null($options['select_usrgrps'])){
 						if(!isset($result[$user['userid']]['usrgrps']))
 							$result[$user['userid']]['usrgrps'] = array();
 
-						$result[$user['userid']]['usrgrps'][$user['usrgrpid']] = array('usrgrpid' => $user['usrgrpid']);
+						$result[$user['userid']]['usrgrps'][] = array('usrgrpid' => $user['usrgrpid']);
 						unset($user['usrgrpid']);
 					}
 
@@ -320,38 +321,38 @@ class CUser extends CZBXAPI{
 		}
 	}
 
-	public static function login($user){	
+	public static function login($user){
 		global $USER_DETAILS, $ZBX_LOCALNODEID;
-	
+
 		$name = $user['user'];
 		$passwd = $user['password'];
 		$auth_type = $user['auth_type'];
 
 		$password = md5($passwd);
-	
+
 		$sql = 'SELECT u.userid,u.attempt_failed, u.attempt_clock, u.attempt_ip '.
 				' FROM users u '.
 				' WHERE u.alias='.zbx_dbstr($name);
-	
+
 //SQL to BLOCK attempts
 //					.' AND ( attempt_failed<'.ZBX_LOGIN_ATTEMPTS.
 //							' OR (attempt_failed>'.(ZBX_LOGIN_ATTEMPTS-1).
 //									' AND ('.time().'-attempt_clock)>'.ZBX_LOGIN_BLOCK.'))';
-	
+
 		$login = $attempt = DBfetch(DBselect($sql));
-	
+
 		if(($name!=ZBX_GUEST_USER) && zbx_empty($passwd)){
 			$login = $attempt = false;
 		}
-	
+
 		if($login){
 			if(($login['attempt_failed'] >= ZBX_LOGIN_ATTEMPTS) && ((time() - $login['attempt_clock']) < ZBX_LOGIN_BLOCK)){
 				$_REQUEST['message'] = 'Account is blocked for ' . (ZBX_LOGIN_BLOCK - (time() - $login['attempt_clock'])) .' seconds.';
 				return false;
 			}
-	
+
 			DBexecute('UPDATE users SET attempt_clock=' . time() . ' WHERE alias='.zbx_dbstr($name));
-	
+
 			switch(get_user_auth($login['userid'])){
 				case GROUP_GUI_ACCESS_INTERNAL:
 					$auth_type = ZBX_AUTH_INTERNAL;
@@ -361,7 +362,7 @@ class CUser extends CZBXAPI{
 				default:
 					break;
 			}
-	
+
 			switch($auth_type){
 				case ZBX_AUTH_LDAP:
 					$login = self::ldapLogin($user);
@@ -375,17 +376,17 @@ class CUser extends CZBXAPI{
 					$login = true;
 			}
 		}
-	
+
 		if($login){
 			$sql = 'SELECT u.userid,u.alias,u.name,u.surname,u.url,u.refresh,u.passwd '.
 						' FROM users u, users_groups ug, usrgrp g '.
 						' WHERE u.alias='.zbx_dbstr($name).
 							((ZBX_AUTH_INTERNAL==$auth_type)?' AND u.passwd='.zbx_dbstr($password):'').
 							' AND '.DBin_node('u.userid', $ZBX_LOCALNODEID);
-	
+
 			$login = $user = DBfetch(DBselect($sql));
 		}
-	
+
 /* update internal pass if it's different
 	if($login && ($row['passwd']!=$password) && (ZBX_AUTH_INTERNAL!=$auth_type)){
 		DBexecute('UPDATE users SET passwd='.zbx_dbstr($password).' WHERE userid='.$row['userid']);
@@ -394,26 +395,26 @@ class CUser extends CZBXAPI{
 		if($login){
 			$login = (check_perm2login($user['userid']) && check_perm2system($user['userid']));
 		}
-	
+
 		if($login){
 			$sessionid = zbx_session_start($user['userid'], $name, $password);
-	
+
 			add_audit(AUDIT_ACTION_LOGIN,AUDIT_RESOURCE_USER,'Correct login ['.$name.']');
-	
+
 			if(empty($user['url'])){
-				$user['url'] = get_profile('web.menu.view.last','index.php');
+				$user['url'] = CProfile::get('web.menu.view.last','index.php');
 			}
-	
-	
+
+
 			$USER_DETAILS = $user;
 			$login = $sessionid;
 		}
 		else{
 			$user = NULL;
-	
+
 			$_REQUEST['message'] = 'Login name or password is incorrect';
 			add_audit(AUDIT_ACTION_LOGIN,AUDIT_RESOURCE_USER,'Login failed ['.$name.']');
-	
+
 			if($attempt){
 				$ip = (isset($_SERVER['HTTP_X_FORWARDED_FOR']) && !empty($_SERVER['HTTP_X_FORWARDED_FOR']))?$_SERVER['HTTP_X_FORWARDED_FOR']:$_SERVER['REMOTE_ADDR'];
 				$attempt['attempt_failed']++;
@@ -424,7 +425,7 @@ class CUser extends CZBXAPI{
 				DBexecute($sql);
 			}
 		}
-	
+
 	return $login;
 	}
 
@@ -432,26 +433,26 @@ class CUser extends CZBXAPI{
 		$name = $user['user'];
 		$passwd = $user['password'];
 		$cnf = isset($user['cnf'])?$user['cnf']:null;
-		
+
 		if(is_null($cnf)){
 			$config = select_config();
 			foreach($config as $id => $value){
-				if(strpos($id,'ldap_') !== false){
+				if(zbx_strpos($id,'ldap_') !== false){
 					$cnf[str_replace('ldap_','',$id)] = $config[$id];
 				}
 			}
 		}
-	
+
 		if(!function_exists('ldap_connect')){
 			info('Probably php-ldap module is missing.');
 			return false;
 		}
-	
+
 		$ldap = new CLdap($cnf);
 		$ldap->connect();
-	
+
 		$result = $ldap->checkPass($name,$passwd);
-	
+
 	return $result;
 	}
 
@@ -477,10 +478,10 @@ class CUser extends CZBXAPI{
 		global	$ZBX_NODES;
 
 		$sessionid = is_null($user)?null:$user['sessionid'];
-	
+
 		$USER_DETAILS = NULL;
 		$login = FALSE;
-	
+
 		if(!is_null($sessionid)){
 			$sql = 'SELECT u.*,s.* '.
 					' FROM sessions s,users u'.
@@ -489,9 +490,9 @@ class CUser extends CZBXAPI{
 						' AND s.userid=u.userid'.
 						' AND ((s.lastaccess+u.autologout>'.time().') OR (u.autologout=0))'.
 						' AND '.DBin_node('u.userid', $ZBX_LOCALNODEID);
-	
+
 			$login = $USER_DETAILS = DBfetch(DBselect($sql));
-	
+
 			if(!$USER_DETAILS){
 				$incorrect_session = true;
 	}
@@ -503,11 +504,11 @@ class CUser extends CZBXAPI{
 							' on ',
 							bold(date('d.m.Y H:i',$login['attempt_clock'])),
 							'.')));
-	
+
 				DBexecute('UPDATE users SET attempt_failed=0 WHERE userid='.$login['userid']);
 			}
 		}
-	
+
 		if(!$USER_DETAILS && !isset($_SERVER['PHP_AUTH_USER'])){
 			$sql = 'SELECT u.* '.
 					' FROM users u '.
@@ -521,15 +522,15 @@ class CUser extends CZBXAPI{
 				$sessionid = zbx_session_start($USER_DETAILS['userid'], ZBX_GUEST_USER, '');
 			}
 		}
-	
+
 		if($login){
 			$login = (check_perm2login($USER_DETAILS['userid']) && check_perm2system($USER_DETAILS['userid']));
 		}
-	
+
 		if(!$login){
 			$USER_DETAILS = NULL;
 		}
-	
+
 		if($login && $sessionid && !isset($incorrect_session)){
 			zbx_setcookie('zbx_sessionid',$sessionid,$USER_DETAILS['autologin']?(time()+86400*31):0);	//1 month
 			DBexecute('UPDATE sessions SET lastaccess='.time().' WHERE sessionid='.zbx_dbstr($sessionid));
@@ -539,7 +540,7 @@ class CUser extends CZBXAPI{
 			DBexecute('UPDATE sessions SET status='.ZBX_SESSION_PASSIVE.' WHERE sessionid='.zbx_dbstr($sessionid));
 			unset($sessionid);
 		}
-	
+
 		if($USER_DETAILS){
 //		$USER_DETAILS['node'] = DBfetch(DBselect('SELECT * FROM nodes WHERE nodeid='.id2nodeid($USER_DETAILS['userid'])));
 			if(isset($ZBX_NODES[$ZBX_LOCALNODEID])){
@@ -550,7 +551,7 @@ class CUser extends CZBXAPI{
 				$USER_DETAILS['node']['name'] = '- unknown -';
 				$USER_DETAILS['node']['nodeid'] = $ZBX_LOCALNODEID;
 			}
-	
+
 			$USER_DETAILS['debug_mode'] = get_user_debug_mode($USER_DETAILS['userid']);
 		}
 		else{
@@ -563,12 +564,12 @@ class CUser extends CZBXAPI{
 					'name'	=>'- unknown -',
 					'nodeid'=>0));
 		}
-	
+
 		$userip = (isset($_SERVER['HTTP_X_FORWARDED_FOR']) && !empty($_SERVER['HTTP_X_FORWARDED_FOR']))?$_SERVER['HTTP_X_FORWARDED_FOR']:$_SERVER['REMOTE_ADDR'];
 		$USER_DETAILS['userip'] = $userip;
-	
+
 		if(!$login || isset($incorrect_session) || isset($missed_user_guest)){
-	
+
 			if(isset($incorrect_session))	$message = 'Session terminated, please re-login!';
 			else if(isset($missed_user_guest)){
 				$row = DBfetch(DBselect('SELECT count(u.userid) as user_cnt FROM users u'));
@@ -576,12 +577,12 @@ class CUser extends CZBXAPI{
 					$message = 'Table users is empty. Possible database corruption.';
 				}
 			}
-	
+
 			if(!isset($_REQUEST['message']) && isset($message)) $_REQUEST['message'] = $message;
-	
+
 		return false;
 		}
-	
+
 	return true;
 	}
 
@@ -647,7 +648,7 @@ class CUser extends CZBXAPI{
  * @param string $users['user_medias']['period']
  * @return array|boolean
  */
-	public static function add($users){
+	public static function create($users){
 		global $USER_DETAILS;
 		$result = false;
 		$errors = array();
@@ -739,8 +740,7 @@ class CUser extends CZBXAPI{
 		}
 		$result = self::EndTransaction($result, __METHOD__);
 		if($result){
-			$upd_users = self::get(array('userids' => $userids, 'extendoutput' => 1));
-			return $upd_users;
+			return array('userids' => $userids);
 		}
 		else{
 			self::setMethodErrors(__METHOD__, $errors);
@@ -793,10 +793,12 @@ class CUser extends CZBXAPI{
 		$users = zbx_toArray($users);
 		$userids = zbx_objectValues($users, 'userid');
 
-		$upd_users = self::get(array(
+		$options = array(
 			'userids' => zbx_objectValues($users, 'userid'),
 			'extendoutput' => 1,
-			'preservekeys' => 1));
+			'preservekeys' => 1
+		);
+		$upd_users = self::get($options);
 		foreach($users as $gnum => $user){
 			//add_audit(AUDIT_ACTION_DELETE, AUDIT_RESOURCE_USER, 'User ['.$user['alias'].']');
 		}
@@ -809,6 +811,14 @@ class CUser extends CZBXAPI{
 
 		foreach($users as $unum => $user){
 			$user_db_fields = $upd_users[$user['userid']];
+
+// check if we change guest user
+			if(($user_db_fields['alias'] == ZBX_GUEST_USER) && isset($user['alias']) && ($user['alias'] != ZBX_GUEST_USER)){
+				$errors[] = array('errno' => ZBX_API_ERROR_PARAMETERS, 'error' => 'Cannot rename guest user');
+				$result = false;
+				break;
+			}
+
 
 // unset if not changed passwd
 			if(isset($user['passwd']) && !is_null($user['passwd'])){
@@ -919,8 +929,7 @@ class CUser extends CZBXAPI{
 		$result = self::EndTransaction($result, __METHOD__);
 
 		if($result){
-			$upd_users = self::get(array('userids' => $userids, 'extendoutput' => 1, 'nopermissions' => 1));
-			return $upd_users;
+			return array('userids' => $userids);
 		}
 		else{
 			self::setMethodErrors(__METHOD__, $errors);
@@ -1015,8 +1024,8 @@ class CUser extends CZBXAPI{
 
 		if($result){
 			$options = array(
-					'nodeids' => id2nodeid($USER_DETAILS['userid']), 
-					'userids' => $USER_DETAILS['userid'], 
+					'nodeids' => id2nodeid($USER_DETAILS['userid']),
+					'userids' => $USER_DETAILS['userid'],
 					'extendoutput' => 1
 				);
 
@@ -1052,12 +1061,14 @@ class CUser extends CZBXAPI{
 
 		$users = zbx_toArray($users);
 		$userids = array();
-		$result = false;
+		$result = true;
 
-		$del_users = self::get(array(
+		$options = array(
 			'userids'=>zbx_objectValues($users, 'userid'),
 			'extendoutput'=>1,
-			'preservekeys'=>1));
+			'preservekeys'=>1
+		);
+		$del_users = self::get($options);
 
 		foreach($del_users as $gnum => $user){
 			if(bccomp($USER_DETAILS['userid'], $user['userid']) == 0){
@@ -1066,7 +1077,7 @@ class CUser extends CZBXAPI{
 			}
 
 			if($del_users[$user['userid']]['alias'] == ZBX_GUEST_USER){
-				$errors[] = array('errno' => ZBX_API_ERROR_PARAMETERS, 'error' => S_CANNOT_DELETE_USER.'[ '.ZBX_GUEST_USER.' ]');
+				$errors[] = array('errno' => ZBX_API_ERROR_PARAMETERS, 'error' => S_CANNOT_DELETE_USER.' [ '.ZBX_GUEST_USER.' ]');
 				$result = false;
 			}
 
@@ -1075,7 +1086,7 @@ class CUser extends CZBXAPI{
 		}
 
 		self::BeginTransaction(__METHOD__);
-		if(!empty($userids)){
+		if(!empty($userids) && $result){
 			$result = DBexecute('DELETE FROM operations WHERE object='.OPERATION_OBJECT_USER.' AND '.DBcondition('objectid', $userids));
 			$result = DBexecute('DELETE FROM media WHERE '.DBcondition('userid', $userids));
 			$result = DBexecute('DELETE FROM profiles WHERE '.DBcondition('userid', $userids));
@@ -1086,7 +1097,7 @@ class CUser extends CZBXAPI{
 		$result = self::EndTransaction($result, __METHOD__);
 
 		if($result){
-			return zbx_cleanHashes($del_users);
+			return array('userids' => $userids);
 		}
 		else{
 			self::setMethodErrors(__METHOD__, $errors);
@@ -1114,9 +1125,9 @@ class CUser extends CZBXAPI{
  */
 	public static function addMedia($media_data){
 		global $USER_DETAILS;
-		
+
 		$result = true;
-		
+
 		$medias = zbx_toArray($media_data['medias']);
 		$users = zbx_toArray($media_data['users']);
 
@@ -1125,9 +1136,22 @@ class CUser extends CZBXAPI{
 			return false;
 		}
 
-		foreach($users as $user){
-			foreach($medias as $media){
-				$result = add_media( $user['userid'], $media['mediatypeid'], $media['sendto'], $media['severity'], $media['active'], $media['period']);
+		foreach($users as $unum => $user){
+			foreach($medias as $mnum => $media){
+					if(!validate_period($media['period'])){
+						self::setError(__METHOD__, ZBX_API_ERROR_PARAMETERS, 'Incorrect time period');
+						return false;
+					}
+
+					$mediaid = get_dbid('media','mediaid');
+
+					$sql='INSERT INTO media (mediaid,userid,mediatypeid,sendto,active,severity,period) '.
+							' VALUES ('.$mediaid.','.$user['userid'].','.$media['mediatypeid'].','.
+										zbx_dbstr($media['sendto']).','.$media['active'].','.$media['severity'].','.
+										zbx_dbstr($media['period']).')';
+
+					$result = DBexecute($sql);
+
 				if(!$result) break 2;
 			}
 		}
@@ -1204,7 +1228,7 @@ class CUser extends CZBXAPI{
 
 		$result = false;
 		$transaction = false;
-		
+
 		$new_medias = zbx_toArray($media_data['medias']);
 		$users = zbx_toArray($media_data['users']);
 
@@ -1228,7 +1252,7 @@ class CUser extends CZBXAPI{
 
 			foreach($new_medias as $mnum => $media){
 				if(!isset($media['mediaid'])) continue;
-				
+
 				if(isset($del_medias[$media['mediaid']])){
 					$upd_medias[$media['mediaid']] = $new_medias[$mnum];
 				}
@@ -1241,7 +1265,7 @@ class CUser extends CZBXAPI{
 			if(!empty($del_medias)){
 				$result = self::deleteMedia($del_medias);
 				if(!$result){
-					throw new APIException(ZBX_API_ERROR_PARAMETERS, 'Cant delete user medias');
+					throw new APIException(ZBX_API_ERROR_PARAMETERS, 'Can\'t delete user medias');
 				}
 			}
 
@@ -1260,15 +1284,15 @@ class CUser extends CZBXAPI{
 						' WHERE mediaid='.$media['mediaid'];
 				$result = DBexecute($sql);
 				if(!$result){
-					throw new APIException(ZBX_API_ERROR_PARAMETERS, 'Cant update user media');
+					throw new APIException(ZBX_API_ERROR_PARAMETERS, 'Can\'t update user media');
 				}
 			}
-						
+
 // CREATE
 			if(!empty($new_medias)){
 				$result = self::addMedia(array('users' => $users, 'medias' => $new_medias));
 				if(!$result){
-					throw new APIException(ZBX_API_ERROR_PARAMETERS, 'Cant insert user media');
+					throw new APIException(ZBX_API_ERROR_PARAMETERS, 'Can\'t insert user media');
 				}
 			}
